BudgetHistoric

+++ VaccineC/src/app/services/budgets-products-dispatcher.service.ts
@@ -37,12 +37,12 @@ export class BudgetsProductsDispatcherService {
         return this.httpClient.put(`${baseURL}/${id}/Update`, data);
     }
 
-    public repeatOnDemand(id: string, numberOfTimes: number, repeatBorrower: boolean): Observable<any> {
-        return this.httpClient.post(`${baseURL}/${id}/${numberOfTimes}/${repeatBorrower}/RepeatOnDemand`, id);
+    public repeatOnDemand(id: string, numberOfTimes: number, repeatBorrower: boolean, userId: string): Observable<any> {
+        return this.httpClient.post(`${baseURL}/${id}/${numberOfTimes}/${repeatBorrower}/${userId}/RepeatOnDemand`, id);
     }
 
-    public deleteBudgetProduct(id: string): Observable<any> {
-        return this.httpClient.delete(`${baseURL}/${id}/Delete`);
+    public deleteBudgetProduct(id: string, userId: string): Observable<any> {
+        return this.httpClient.delete(`${baseURL}/${id}/${userId}/Delete`);
     }
 
 }
